+++ src/crypto/hashing.rs
@@ -1,3 +1,5 @@
+use crate::crypto::ethereum::pubkey_error;
+use cosmwasm_std::StdError;
 use ripemd::Ripemd160;
 use sha2::{Digest, Sha256};
 use tiny_keccak::Hasher;
@@ -30,4 +32,25 @@ pub fn ripemd160(data: &[u8]) -> Vec<u8> {
     let mut ripemd160 = Ripemd160::new();
     ripemd160.update(data);
     ripemd160.finalize().to_vec()
+}
+
+pub fn compress_key(uncompressed_pubkey_bytes: &[u8]) -> Result<Vec<u8>, StdError> {
+    if uncompressed_pubkey_bytes.len() != 64 {
+        return Err(pubkey_error(&"Wrong len"));
+    }
+
+    // The first byte is the prefix, followed by 32 bytes for X and 32 bytes for Y
+    let x_bytes = &uncompressed_pubkey_bytes[0..32];
+    let y_bytes = &uncompressed_pubkey_bytes[32..64];
+
+    // Determine if Y is even or odd for the prefix
+    // Y's last byte's least significant bit determines its evenness or oddness
+    let prefix_byte = if y_bytes[31] & 1 == 0 { 0x02 } else { 0x03 };
+
+    // Create the compressed public key
+    let mut compressed_pubkey: Vec<u8> = Vec::with_capacity(33);
+    compressed_pubkey.push(prefix_byte);
+    compressed_pubkey.extend_from_slice(x_bytes);
+
+    Ok(compressed_pubkey)
 }
